infrastructure/cdk/pyproject.toml
[project]
name = "sbir-analytics-infrastructure"
version = "0.1.0"
description = "AWS CDK infrastructure for SBIR ETL pipeline"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
